- Add an opt-in `clang-tidy` CMake target (`cmake/clang_tidy.cmake`) that runs clang-tidy over `tlv` (C) and `tlv++` (C++) against the build's `compile_commands.json`; like `format`/`format-check`/`cppcheck`, it's skipped with a warning (never breaking a regular build) if clang-tidy isn't found or `CMAKE_EXPORT_COMPILE_COMMANDS` wasn't enabled. Check selection lives in a root [.clang-tidy](.clang-tidy) baseline (`clang-diagnostic-*`, `clang-analyzer-*`, `bugprone-*`, `performance-*`, `portability-*`, `readability-*`, with a handful of checks excluded that are either pure style preferences conflicting with this codebase's conventions or would need broad, correctness-unrelated rewrites) plus [tlv/.clang-tidy](tlv/.clang-tidy) and [tlv++/.clang-tidy](tlv++/.clang-tidy), which clang-tidy layers on top per-directory to keep C++-only checks (a curated `modernize-*` subset) off of `tlv`. Since `tlv++` is header-only, its coverage runs through `tests/unit/tlv++/src/test_tlvpp.cpp`, which already exercises every `tlv++` header's templates with real instantiations, rather than a synthetic `#include`-only translation unit that would compile without ever instantiating them. (#120) - Add a `clang-tidy` job to [static-analysis.yml](.github/workflows/static-analysis.yml), configuring the project with CMake + Ninja and Clang 18 and building the `clang-tidy` target, alongside the existing pre-commit formatting and cppcheck jobs. The job only needs to configure (not build) the project, since `compile_commands.json` is written at CMake generate time, and always writes a findings table (file, line, severity, message) to the job summary, independent of whether the job passes or fails. (#120) - Fix the clang-tidy findings the new check surfaced across `tlv` and `tlv++` (unparenthesized macro parameters, a redundant cast, non-uppercase integer literal suffixes, several narrow-then-widened multiplications made explicit, two uninitialized locals left garbage on an unverified cross-file success contract, a public schema-dictionary lookup missing a defense-in-depth bounds check, a redundant header include, and a template polyfill's move constructor missing a conditional `noexcept`) so the check starts green; a few remaining Clang Static Analyzer findings are genuine false positives (through indirect function-pointer calls or a recursive arena-write chain the analyzer can't fully fold) and are suppressed with a `NOLINT`/`NOLINTBEGIN`-`NOLINTEND` comment explaining why. (#120) - Document how to run clang-tidy locally in [CONTRIBUTING.md](CONTRIBUTING.md#static-analysis-clang-tidy). (#120)

+++ tlv/src/codec/emv.c
@@ -213,7 +213,7 @@ tlv_codec_result_t emv_value_encode(const void* context, const void* value, size
             tlv_emv_date_t date;
             EMV_LOAD(date);
             if (!valid_date(date)) return TLV_CODEC_ERR_INVALID_VALUE;
-            number = date.year * UINT64_C(10000) + date.month * 100 + date.day;
+            number = date.year * UINT64_C(10000) + (uint64_t)date.month * 100 + date.day;
             write_number(number, 6, bytes, count);
             break;
         }
@@ -222,7 +222,7 @@ tlv_codec_result_t emv_value_encode(const void* context, const void* value, size
             EMV_LOAD(time);
             if (time.hour > 23 || time.minute > 59 || time.second > 59)
                 return TLV_CODEC_ERR_INVALID_VALUE;
-            number = time.hour * UINT64_C(10000) + time.minute * 100 + time.second;
+            number = time.hour * UINT64_C(10000) + (uint64_t)time.minute * 100 + time.second;
             write_number(number, 6, bytes, count);
             break;
         }
